refactor: move fetch methods to separate files
